added inttypes.h and changed count_digits to intmax_t

+++ ft_count_digits_base.c
@@ -6,13 +6,13 @@
 /*   By: cpireyre <cpireyre@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2018/06/13 11:54:06 by cpireyre          #+#    #+#             */
-/*   Updated: 2018/06/14 09:22:35 by cpireyre         ###   ########.fr       */
+/*   Updated: 2018/06/16 10:27:47 by cpireyre         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-int		ft_count_digits_base(long long n, int base)
+int		ft_count_digits_base(intmax_t n, int base)
 {
 	int	count;
